feat: rust problems
run solutions test, general test fn
