--- problems/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "problems"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.79.0"
-authors = ["benhao"]
-description = "LeetCode Solutions in Rust"
-readme = "../README.md"
-
-[features]
-solution = []
-
-[dependencies]
-serde_json = "1.0"
-rand = "0.8.4"
-regex = "1.10.5"
-library = { path = "../rust/library", features = ["model"]}
-
-[lib]
-name = "solution"
-path = "problems_2/solution.rs"
