+++ problems/problems_1/solution.rs
@@ -16,7 +16,7 @@ impl Solution {
     }
 }
 
-#[cfg(feature = "solution")]
+#[cfg(feature = "solution_1")]
 pub fn solve(input_string: String) -> Value {
     let input_values: Vec<String> = input_string.split('\n').map(|x| x.to_string()).collect();
     let nums: Vec<i32> = serde_json::from_str(&input_values[0]).expect("Failed to parse input");
