+++ problems/problems_2/solution.rs
@@ -42,7 +42,7 @@ impl Solution {
     }
 }
 
-#[cfg(feature = "solution")]
+#[cfg(feature = "solution_2")]
 pub fn solve(input_string: String) -> Value {
     let input_values: Vec<String> = input_string.split('\n').map(|x| x.to_string()).collect();
     let nums0: Vec<i32> = serde_json::from_str(&input_values[0]).expect("Failed to parse input");
