+++ rust/test_executor/Cargo.toml
@@ -8,9 +8,7 @@ description = "LeetCode Rust Test Main"
 readme = "../README.md"
 
 [features]
-timer = []
-test_case = []
+run_test = []
 
 [dependencies]
 serde_json = "1.0"
-problems = { path = "../../problems", features = ["solution"] }
